requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
opencv-python>=4.7.0
Pillow>=9.0.0
scikit-learn>=1.2.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
tensorboard>=2.12.0
matplotlib>=3.6.0
seaborn>=0.12.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3
nltk>=3.8
sacrebleu>=2.3.0
rouge-score>=0.1.2
tqdm>=4.64.0
rich>=13.0.0
click>=8.1.0
pyyaml>=6.0
pytest>=7.2.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
transformers>=4.25.0
timm>=0.6.0
